+++ starkinfra/creditnote/__init__.py
@@ -6,4 +6,4 @@ from . import invoice
 from .invoice.__invoice import Invoice
 from .invoice.__discount import Discount
 from .invoice.__description import Description
-from .__creditnote import create, get, query, page, cancel, pdf
+from .__creditnote import create, get, query, page, cancel, pdf, payment